+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix user config CM mapping for bundle apps.
+
 ### Added
 
 - Read app dependencies from Release CR to avoid deadlock installing apps in new clusters.